+++ components/NotificationBadge.tsx
@@ -4,7 +4,7 @@ import { View, Text, StyleSheet } from 'react-native';
 import NotificationService from '../services/notification';
 import { useTheme } from '../contexts/ThemeContext';
 import { useAuth } from '@/features/auth/AuthContext';
-import { spacing } from '@/constants/tokens';
+import { spacing } from 'src/shared/ui/tokens';
 
 interface NotificationBadgeProps {
   size?: number;
