+++ components/NotificationItem.tsx
@@ -3,7 +3,7 @@ import { View, Text, StyleSheet, TouchableOpacity, Platform } from 'react-native
 import { Package, Tag, MessageCircle, Info, CircleCheck as CheckCircle, CircleAlert as AlertCircle } from 'lucide-react-native';
 import { Notification } from '../types';
 import { useTheme } from '../contexts/ThemeContext';
-import { spacing, radius, shadows } from '@/constants/tokens';
+import { spacing, radius, shadows } from 'src/shared/ui/tokens';
 import { typography } from '@/constants/styles';
 
 interface NotificationItemProps {
