+++ components/NotificationPopup.tsx
@@ -10,7 +10,7 @@ import {
 } from 'react-native';
 import { X } from 'lucide-react-native';
 import { useTheme } from '../contexts/ThemeContext';
-import { spacing, zIndex } from '@/constants/tokens';
+import { spacing, zIndex } from 'src/shared/ui/tokens';
 
 interface NotificationPopupProps {
   title: string;
