+++ src/shared/ui/tokens.ts
@@ -1,0 +1,2 @@
+export { spacing, radius, colors, zIndex, shadows, tokens, lightTokens } from '@/constants/tokens';
+export type { Tokens } from '@/constants/tokens';
